# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SchoolLedger Kenya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fees/structures:
    get:
      operationId: listFeeStructures
      tags: [fees]
      summary: List fee structures
      parameters:
        - name: academicYear
          in: query
          schema:
            type: string
        - name: term
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fee structures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeStructure"
    post:
      operationId: createFeeStructure
      tags: [fees]
      summary: Create a fee structure for a class/term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeStructureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStructure"
  /fees/structures/{id}:
    patch:
      operationId: updateFeeStructure
      tags: [fees]
      summary: Update a fee structure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeStructureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStructure"
    delete:
      operationId: deleteFeeStructure
      tags: [fees]
      summary: Delete fee structure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /fees/defaulters:
    get:
      operationId: listDefaulters
      tags: [fees]
      summary: Students with outstanding fee balances for current term
      parameters:
        - name: class
          in: query
          schema:
            type: string
        - name: minBalance
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Defaulters list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Defaulter"

  # ── Payments ───────────────────────────────────────────────────────────────
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all statutory fund accounts
      responses:
        "200":
          description: Fund accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FundAccount"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a statutory fund account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundAccount"
  /account-transactions:
    get:
      operationId: listAccountTransactions
      tags: [accounts]
      summary: Get transaction ledger for a fund account
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Account transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountTransaction"

  # ── Capitation ─────────────────────────────────────────────────────────────
  /capitation/disbursements:
    get:
      operationId: listCapitationDisbursements
      tags: [capitation]
      summary: List government capitation disbursements received
      responses:
        "200":
          description: Capitation disbursements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CapitationDisbursement"
    post:
      operationId: recordCapitationDisbursement
      tags: [capitation]
      summary: Record a new capitation disbursement from MoE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapitationDisbursementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapitationDisbursement"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expense vouchers
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Expense vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: voidExpense
      tags: [expenses]
      summary: Void an expense (audit trail preserved)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voided expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── Reports ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FeeStructure:
      type: object
      required: [id, academicYear, term, class, totalAmount, breakdown, createdAt]
      properties:
        id:
          type: integer
        academicYear:
          type: string
        term:
          type: integer
        class:
          type: string
        totalAmount:
          type: number
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/FeeLineItem"
        createdAt:
          type: string
          format: date-time
    FeeLineItem:
      type: object
      required: [description, amount, voteCategory]
      properties:
        description:
          type: string
        amount:
          type: number
        voteCategory:
          type: string
          enum: [tuition, operations, bom, capitation]
    FeeStructureInput:
      type: object
      required: [academicYear, term, class, breakdown]
      properties:
        academicYear:
          type: string
        term:
          type: integer
        class:
          type: string
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/FeeLineItem"
    FeeStructureUpdate:
      type: object
      properties:
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/FeeLineItem"
    Defaulter:
      type: object
      required: [studentId, admissionNumber, fullName, class, feeExpected, totalPaid, balance]
      properties:
        studentId:
          type: integer
        admissionNumber:
          type: string
        fullName:
          type: string
        class:
          type: string
        guardianPhone:
          type: ["string", "null"]
        feeExpected:
          type: number
        totalPaid:
          type: number
        balance:
          type: number

    # Payments
...
    FundAccount:
      type: object
      required: [id, name, accountType, description, currentBalance, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        accountType:
          type: string
          enum: [tuition, operations, bom, capitation]
        description:
          type: string
        currentBalance:
          type: number
        createdAt:
          type: string
          format: date-time
    FundAccountInput:
      type: object
      required: [name, accountType, description]
      properties:
        name:
          type: string
        accountType:
          type: string
          enum: [tuition, operations, bom, capitation]
        description:
          type: string
    AccountTransaction:
      type: object
      required: [id, accountId, type, amount, balance, description, referenceType, referenceId, transactionDate]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        balance:
          type: number
        description:
          type: string
        referenceType:
          type: string
          enum: [payment, expense, capitation, adjustment]
        referenceId:
          type: integer
        transactionDate:
          type: string
          format: date-time

    # Capitation
    CapitationDisbursement:
      type: object
      required: [id, academicYear, term, amount, disbursementDate, moeReference, studentCount, perStudentRate, notes, createdAt]
      properties:
        id:
          type: integer
        academicYear:
          type: string
        term:
          type: integer
        amount:
          type: number
        disbursementDate:
          type: string
          format: date
        moeReference:
          type: string
        studentCount:
          type: integer
        perStudentRate:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CapitationDisbursementInput:
      type: object
      required: [academicYear, term, amount, disbursementDate, moeReference, studentCount]
      properties:
        academicYear:
          type: string
        term:
          type: integer
        amount:
          type: number
        disbursementDate:
          type: string
          format: date
        moeReference:
          type: string
        studentCount:
          type: integer
        notes:
          type: string
...
    Expense:
      type: object
      required: [id, accountId, amount, description, category, voucherNumber, expenseDate, approvedBy, status, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        approvedBy:
          type: string
        supplierName:
          type: ["string", "null"]
        status:
          type: string
          enum: [approved, void]
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [accountId, amount, description, category, voucherNumber, expenseDate, approvedBy]
      properties:
        accountId:
          type: integer
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        approvedBy:
          type: string
        supplierName:
          type: string

    # Reports
...