# lib/api-zod/src/generated/types/defaulter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SchoolLedger Kenya API
 * OpenAPI spec version: 0.1.0
 */

export interface Defaulter {
  studentId: number;
  admissionNumber: string;
  fullName: string;
  class: string;
  /** @nullable */
  guardianPhone?: string | null;
  feeExpected: number;
  totalPaid: number;
  balance: number;
}