# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SchoolLedger Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseStatus } from './expenseStatus';

export interface Expense {
  id: number;
  accountId: number;
  amount: number;
  description: string;
  category: string;
  voucherNumber: string;
  expenseDate: Date;
  approvedBy: string;
  /** @nullable */
  supplierName?: string | null;
  status: ExpenseStatus;
  createdAt: Date;
}