# lib/api-zod/src/generated/types/feeStructure.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SchoolLedger Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { FeeLineItem } from './feeLineItem';

export interface FeeStructure {
  id: number;
  academicYear: string;
  term: number;
  class: string;
  totalAmount: number;
  breakdown: FeeLineItem[];
  createdAt: Date;
}